requirements.txt
flask
flask-login
flask-sqlalchemy
Flask-WTF
intasend-python
gunicorn
